requirements.txt
openai>=1.10.0
pillow>=9.0.0
tqdm>=4.62.0
requests>=2.27.0
